Fix NPE in clean-up: Test existence of reports dir
The directory for client error reports is not always present, as it is
created on demand by tcos_exception_hook (in tcos-libs).

Clean up old report files only if directory exists (instead of blindly
provoking a NPE).

+++ runtime/standalone/src/main/java/org/openthinclient/manager/standalone/config/ManagerHomeCleaner.java
@@ -55,9 +55,11 @@ public class ManagerHomeCleaner {
    */
   private void cleanReports(long notBefore) {
     File reportsDir = managerHomePath.resolve(REPORTS_PATH).toFile();
-    for(File subdir : reportsDir.listFiles(File::isDirectory)) {
-      for(File file : subdir.listFiles(file -> file.lastModified() < notBefore)) {
-        deleteFile(file);
+    if(reportsDir.isDirectory()) {
+      for(File subdir : reportsDir.listFiles(File::isDirectory)) {
+        for(File file : subdir.listFiles(file -> file.lastModified() < notBefore)) {
+          deleteFile(file);
+        }
       }
     }
   }
